third-party/requirements.txt
pymongo == 4.3.3
requests == 2.28.2
python-dotenv == 1.0.0
APScheduler == 3.10.1
Flask == 2.2.3
Flask-RESTful == 0.3.9